Optimize includes of Exception.h
