+++ programs/keeper/keeper_main.cpp
@@ -19,8 +19,9 @@
 #include <Common/getHashOfLoadedBinary.h>
 #include <Common/IO.h>
 
-#include <base/phdr_cache.h>
 #include <base/coverage.h>
+#include <base/phdr_cache.h>
+#include <base/scope_guard.h>
 
 
 int mainEntryClickHouseKeeper(int argc, char ** argv);
