+++ programs/su/su.cpp
@@ -1,6 +1,8 @@
 #include <Common/Exception.h>
 #include <IO/ReadHelpers.h>
 #include <fmt/format.h>
+
+#include <iostream>
 #include <vector>
 
 #include <sys/types.h>
