+++ src/Common/Exception.cpp
@@ -56,7 +56,6 @@ void abortOnFailedAssertion(const String & description)
 
 bool terminate_on_any_exception = false;
 static int terminate_status_code = 128 + SIGABRT;
-thread_local bool update_error_statistics = true;
 std::function<void(const std::string & msg, int code, bool remote, const Exception::FramePointers & trace)> Exception::callback = {};
 
 /// - Aborts the process if error code is LOGICAL_ERROR.
@@ -74,9 +73,6 @@ static void handle_error_code(const std::string & msg, int code, bool remote, co
 
     if (Exception::callback)
         Exception::callback(msg, code, remote, trace);
-
-    if (!update_error_statistics) [[unlikely]]
-        return;
 
     ErrorCodes::increment(code, remote, msg, trace);
 }
@@ -293,7 +289,7 @@ void tryLogCurrentException(const char * log_name, const std::string & start_of_
     LockMemoryExceptionInThread lock_memory_tracker(VariableContext::Global);
 
     /// getLogger can allocate memory too
-    auto logger = getLogger(log_name);
+    auto logger = getLogger(String{log_name});
     tryLogCurrentExceptionImpl(logger.get(), start_of_message, level);
 }
