+++ src/Common/checkStackSize.cpp
@@ -1,5 +1,6 @@
 #include <base/getThreadId.h>
 #include <base/defines.h> /// THREAD_SANITIZER
+#include <base/scope_guard.h>
 #include <Common/checkStackSize.h>
 #include <Common/Exception.h>
 #include <Common/Fiber.h>
