+++ src/IO/examples/limit_read_buffer2.cpp
@@ -1,10 +1,10 @@
-#include <sstream>
-
 #include <IO/LimitReadBuffer.h>
 #include <IO/copyData.h>
 #include <IO/WriteBufferFromString.h>
 #include <IO/ReadHelpers.h>
 
+#include <iostream>
+#include <sstream>
 
 namespace DB
 {
