+++ src/Parsers/ASTAlterQuery.cpp
@@ -2,6 +2,7 @@
 
 #include <Core/ServerSettings.h>
 #include <IO/Operators.h>
+#include <base/scope_guard.h>
 #include <Common/quoteString.h>
